Move link to module
